Fix issue that occurs when 'import' statements appear *between* metadata and the thing the metadata decorates
when this happens the 'import' is visited/emitted first, skipping over the source containing the metadata, and then the metadata is visited along with the thing it decorates and an attempt at emitting it (i.e. commenting it out) happens but the emitter is already past the point in the source where the metadata exists, resulting in incorrect emission.

This fix corrects this issue by looking backwards in the already existing output and affecting the emitted metadata there (because the metadata was 'jumped over' when the 'import' was emitted) instead of trying to emit it again